+++ test/settings/sleep.py
@@ -1,0 +1,43 @@
+"""Datalogger configuration"""
+
+# General settings.
+main = {
+
+    # Measurement intervals in seconds.
+    'interval': {
+
+        # Apply this interval if device is in field mode.
+        'field': 15.0,
+
+        # Apply this interval if device is in maintenance mode.
+        # https://community.hiveeyes.org/t/wartungsmodus-fur-den-terkin-datenlogger/2274
+        'maintenance': 2.0,
+
+    },
+
+    # Whether to use light sleep between measurement cycles.
+    'lightsleep': False,
+
+    # Whether to use deep sleep between measurement cycles.
+    'deepsleep': False,
+
+    # Configure logging.
+    'logging': {
+
+        # Enable or disable logging completely.
+        'enabled': True,
+
+        # Log configuration settings at system startup.
+        'configuration': False,
+
+    },
+
+}
+
+# Networking configuration.
+networking = {
+    'wifi': {
+        # Enable/disable WiFi completely.
+        'enabled': False,
+    },
+}