+++ CHANGES.rst
@@ -142,6 +142,7 @@ Development
 - Skip Mosquitto/Docker integration tests on Windows/WSL
 - Improve test suite re. Mosquitto dependency
 - Add LoRaWAN tests
+- Add tests covering sleep modes and maintenance mode
 
 
 2019-08-19 0.6.0